"0.26.14"

+++ mycotools/mycotools/lib/kontools.py
@@ -77,7 +77,8 @@ def tardir(dir_, rm = True):
         return False
     with tarfile.open(format_path(dir_)[:-1] + '.tar.gz', 'w:gz') as tar:
         tar.add(dir_, arcname = os.path.basename(format_path(dir_)[:-1]))
-    shutil.rmtree(dir_)
+    if rm:
+        shutil.rmtree(dir_)
 
 def untardir(dir_, rm = False, to = None):
     if not to:
@@ -615,7 +616,8 @@ def file2list(file_, types = '', sep = None, col = None, compress = False):
     return data1_list
 
 
-def multisub(args_lists, processes = 1, shell = False, stdout = None, rm = False):
+def multisub(args_lists, processes = 1, shell = False, 
+             verbose = False):
     '''
     Inputs: list of arguments, integer of processes, subprocess `shell` bool
     Outputs: launches and monitors subprocesses and returns list of exit 
@@ -634,85 +636,45 @@ def multisub(args_lists, processes = 1, shell = False, stdout = None, rm = False
         are complete before exiting the function.
     '''
 
-    running, outputs = [ ], []
+    if verbose:
+        stdout, stderr = None, None
+    else:
+        stdout = subprocess.DEVNULL
+        stderr = subprocess.DEVNULL
+
+    running, outputs = [], []
     if processes <= 0:
         processes = 1
 
-    if not stdout:
-        while args_lists:
-            while len( running ) < processes and args_lists:
-                cmd = args_lists[ 0 ]
-                run_temp = subprocess.Popen( cmd , stdout = subprocess.DEVNULL, \
-                    stderr = subprocess.DEVNULL, shell = shell )
-                running.append( [run_temp, cmd ] )
-                del args_lists[ 0 ]
+    while args_lists: # while there are arguments populating arg_lists
+        while len(running) < processes and args_lists: # while not @ max
+        # processes
+            cmd = args_lists[0]
+            run_temp = subprocess.Popen(cmd , stdout = stdout, \
+                stderr = stderr, shell = shell) # run command
+            running.append([run_temp, cmd]) # add command to running
+            del args_lists[0] # delete from argument list to run
 
-            if len( args_lists ) > 0:
-                for index, handle in enumerate(running):
-                    handle[0].poll()
-                    returncode = handle[0].returncode
-                    if returncode is not None:
-                        outputs.append( {
-                            'stdin': handle[1], 'code': returncode
-                            } )
-                        del running[ index ]
-                        break
+        if len(args_lists) > 0: # if there are remaining arguments
+            for index, handle in enumerate(running): # check each status
+                handle[0].poll()
+                returncode = handle[0].returncode
+                if returncode is not None:
+                    outputs.append({
+                        'stdin': handle[1], 'code': returncode
+                        }) # add command and exit code
+                    del running[index] # remove from running
+                    break
 
-            else:
-                while len( running ) > 0:
-                    handle = running[ 0 ]
-                    handle[0].poll()
-                    returncode = handle[0].returncode
-                    if returncode is not None:
-                        outputs.append( {
-                            'stdin': handle[1], 'code': returncode
-                            } )
-                        del running[ 0 ]
+        else: # if there aren't remaining arguments
+            while len(running) > 0: # hold until all commands complete
+                handle = running[0]
+                handle[0].poll()
+                returncode = handle[0].returncode
+                if returncode is not None:
+                    outputs.append({
+                        'stdin': handle[1], 'code': returncode
+                        })
+                    del running[0]
 
-    else:
-        count, outs = 0, []
-        while args_lists:
-            while len( running ) < processes and args_lists:
-                cmd = args_lists[ 0 ][0]
-                running.append([
-                    cmd, args_lists[0][1], 
-                    open(args_lists[0][1], 'wb')
-                    ])
-                run_temp = subprocess.Popen( cmd, stdout = running[-1][2],
-                    stderr = subprocess.DEVNULL, shell = shell )
-                running[-1].append(run_temp)
-                if rm:
-                    running[-1].append(args_lists[0][2])
-                del args_lists[ 0 ]
-
-            if len( args_lists ) > 0:
-                for index, handle in enumerate(running):
-                    handle[3].poll()
-                    returncode = handle[3].returncode
-                    if returncode is not None:
-                        if rm:
-                            os.remove(handle[4])
-                        handle[2].close()
-                        outputs.append( {
-                            'stdin': handle[0], 'code': returncode,
-                            'output': handle[1]
-                            } )
-                        del running[ index ]
-                        break
-
-            else:
-                while len( running ) > 0:
-                    handle = running[ 0 ]
-                    handle[3].poll()
-                    returncode = handle[3].returncode
-                    if returncode is not None:
-                        handle[2].close()
-                        if rm:
-                            os.remove(handle[4])
-                        outputs.append( {
-                            'stdin': handle[0], 'code': returncode,
-                            'output': handle[1]
-                            } )
-                        del running[ 0 ]
-
-    return outputs  
+    return outputs # [{stdin: '', code: int()}]
